+++ src/Models/user/UserTypesHistory.ts
@@ -1,0 +1,7 @@
+import UserTypesHistoryInterface from '../../Interfaces/user/UserTypesHistory';
+
+export default class UserTypesHistory {
+    constructor(parameters: UserTypesHistoryInterface) {
+        Object.assign(this, parameters);
+    }
+}